==> notebooks/2.Data_Cleaning.ipynb <==
## Data Cleaning

### Note:
    - This is not the final stage of the data.
    - This is an interim stage before data is processed for supervised time series model training
    - Will do EDA on this interim stage becuase of readability.

try:
    sc.stop()
except:
    pass

from pyspark import SparkContext
from bigdl.util.common import *


# create sparkcontext with bigdl configuration
sc = SparkContext.getOrCreate(conf=create_spark_conf().setMaster("local[*]"))

from pyspark.sql import SparkSession
spark = SparkSession(sparkContext = sc)

spark.conf.set("spark.sql.session.timeZone", "UTC")

#Gather Arkansas Dataset for cleansing
years = list(range(1998, 2018))
for ix, year in enumerate(years):
    if ix == 0:
        df = spark.read.csv(f"hdfs://localhost:9000/solar_data/raw/Arkansas_{year}.csv", inferSchema=True, header=True)
    else:
        to_append = spark.read.csv(f"hdfs://localhost:9000/solar_data/raw/Arkansas_{year}.csv", inferSchema=True, header=True)
        df = df.union(to_append)

df.show(5)

df.count()

df = df.withColumnRenamed('_c0', 'timestamp')

df.printSchema()

df.select('Fill Flag').groupBy('Fill Flag').count().show()

### Fill Flag Meaning:


0: NaN (no flags raise)   
1: Missing Image 	
2: Low Irradiance 	
3: Exceeds Clearsky 	
4: Missing CLoud Properties (does not mean cloud type is Nan as shown below)  
5: Rayleigh Violation

df.select('Cloud Type').groupBy('Cloud Type').count().show()

### Cloud Type Meaning:

-15: NaN  
0: Clear   
1: Probably Clear 
2: Fog 
3: Water  
4: Super-Cooled Water 
5: Mixed 
6: Opaque Ice  
7: Cirrus  
8: Overlapping  
9: Overshooting  
10: Unknown 	 
11: Dust  
12: Smoke

#40 unique duplicate entries!
df.groupby('timestamp').count().where('count > 1').sort('count', ascending=False).sort('timestamp').show(40)

len(years)
#the number of duplicates = 4 * len(years)

df.where('timestamp == "1998-04-05 03:00:00"').show()
#The hour column doesn't match the timestamp

df.where('timestamp == "2014-03-09 03:00:00"').show()
#The hour column doesn't match the timestamp

#Proof that each 'duplicate' timestamp has unique hour column
tss = df.groupby('timestamp').count().where('count > 1').sort('count', ascending=False).select('timestamp')

tss.count()

dup_rows = df.join(tss,on='timestamp', how='right')
dup_rows.show()

dup_rows.count()

from pyspark.sql.functions import col

dup_rows.groupby('timestamp', 'Hour').count().sort(col('count').desc()).show()

#These hour issues are from daylight savings time!!
#This means that switching to UTC should get rid of the Null entries when trying to convert timestamps

#I'll update the timestamp with the uniquely identifying Year|Month|Day|Hour|Minute columns
from pyspark.sql import functions as F

df = (df.withColumn('Month', F.when(F.length(F.col('Month')) == 1, F.concat(F.lit('0'), F.col('Month'))).otherwise(F.col('Month')))
        .withColumn('Day', F.when(F.length(F.col('Day')) == 1, F.concat(F.lit('0'), F.col('Day'))).otherwise(F.col('Day')))
        .withColumn('Hour', F.when(F.length(F.col('Hour')) == 1, F.concat(F.lit('0'), F.col('Hour'))).otherwise(F.col('Hour')))
        .withColumn('Minute', F.when(F.length(F.col('Minute')) == 1, F.concat(F.lit('0'), F.col('Minute'))).otherwise(F.col('Minute')))
        .withColumn('time', F.to_timestamp(F.concat(*['Year', 'Month', 'Day', 'Hour', 'Minute']), format='yyyyMMddHHmm'))
     )

df = df.drop('timestamp')
df = df.withColumnRenamed('time', 'timestamp')

df.where('timestamp == "2008-03-09 07:30:00"').show()


df.printSchema()

#No more duplicate timestamps.
df.groupby('timestamp').count().where('count > 1').sort('count', ascending=False).sort('timestamp').show()

###Checking for null values
df.toPandas().isnull().sum()

### Now I'll make the above into cleaning functions to run on all my data.
def gather_df(name):
    """Concats list of csvs from 1998 to 2018 for a given name"""
    years = list(range(1998, 2018))
    for ix, year in enumerate(years):
        if ix == 0:
            df = spark.read.csv(f"hdfs://localhost:9000/solar_data/raw/{name}_{year}.csv", inferSchema=True, header=True)
        else:
            to_append = spark.read.csv(f"hdfs://localhost:9000/solar_data/raw/{name}_{year}.csv", inferSchema=True, header=True)
            df = df.union(to_append)
    df = df.withColumnRenamed('_c0', 'timestamp')
    return df

def fix_timestamps(df):
    """Fixes the timestamp on daylight savings time"""
    df = (df.withColumn('Month', F.when(F.length(F.col('Month')) == 1, F.concat(F.lit('0'), F.col('Month'))).otherwise(F.col('Month')))
        .withColumn('Day', F.when(F.length(F.col('Day')) == 1, F.concat(F.lit('0'), F.col('Day'))).otherwise(F.col('Day')))
        .withColumn('Hour', F.when(F.length(F.col('Hour')) == 1, F.concat(F.lit('0'), F.col('Hour'))).otherwise(F.col('Hour')))
        .withColumn('Minute', F.when(F.length(F.col('Minute')) == 1, F.concat(F.lit('0'), F.col('Minute'))).otherwise(F.col('Minute')))
        .withColumn('time', F.to_timestamp(F.concat(*['Year', 'Month', 'Day', 'Hour', 'Minute']), format='yyyyMMddHHmm'))
     )

    df = df.drop('timestamp')
    df = df.withColumnRenamed('time', 'timestamp')
    return df

%cd /home/dyllanjr/Solar_Irradiance_Prediction/data/interim/
names = ['Arkansas', 'Arizona', 'Georgia']
for name in names:
    df = gather_df(name)
    df = fix_timestamps(df)
    df.write.csv(f"/home/dyllanjr/Solar_Irradiance_Prediction/data/interim/{name}_interim.csv", header=True)

os.system('hadoop fs -copyFromLocal /home/dyllanjr/Solar_Irradiance_Prediction/data/interim /solar_data')
