==> notebooks/7.Model_Selection.ipynb <==
try:
    sc.stop()
except:
    pass
from pyspark import SparkContext


sc = SparkContext(appName='ModelSelection').getOrCreate()
from pyspark.sql import SparkSession
spark = SparkSession(sparkContext = sc)
spark.conf.set("spark.sql.session.timeZone", "UTC")

df = spark.read.csv(f"hdfs://localhost:9000/solar_data/processed/Arkansas_1.csv", inferSchema=True, header=True)

from pyspark.ml import Pipeline
from pyspark.ml.regression import RandomForestRegressor, GBTRegressor
from pyspark.ml.tuning import CrossValidator, ParamGridBuilder
from pyspark.ml.evaluation import RegressionEvaluator
from pyspark.ml.feature import VectorAssembler


spark

###Random Forest Regressor
cols = df.columns
cols.remove('sum(GHI)')

#vectorize features
assembler = VectorAssembler(inputCols=cols, outputCol='features')
data = assembler.transform(df)

#train/test split
train, test = data.randomSplit([.8, .2])

##Configuring ML pipeline (might use format later)
rf = RandomForestRegressor(featuresCol='features', labelCol='sum(GHI)')
pipeline = Pipeline(stages=[rf])


paramGrid = (ParamGridBuilder()
                 .addGrid(rf.numTrees, [5, 10, 48]) #Extremely buggy, couldn't tune too many params
                 .build())

crossval = CrossValidator(estimator=pipeline,
                         estimatorParamMaps=paramGrid,
                         evaluator=RegressionEvaluator(labelCol='sum(GHI)'),
                         numFolds=3)

rfmodel = crossval.fit(train)


rf1 = rfmodel.bestModel.stages[0] 

rf1.save("/home/dyllanjr/Solar_Irradiance_Prediction/models/Random_Forest")

evaluator = RegressionEvaluator(labelCol='sum(GHI)')
train_preds = rf1.transform(train)
print("Train RMSE: ", evaluator.evaluate(train_preds))
evaluator = RegressionEvaluator(labelCol='sum(GHI)', metricName='r2')
print("Train r2: ", evaluator.evaluate(train_preds))

evaluator = RegressionEvaluator(labelCol='sum(GHI)')
test_preds = rf1.transform(test)
print("Test RMSE: ", evaluator.evaluate(test_preds))
evaluator = RegressionEvaluator(labelCol='sum(GHI)', metricName='r2')
print("Test r2: ", evaluator.evaluate(test_preds))

###Gradient Boosted Tree Regressor
cols = df.columns
cols.remove('sum(GHI)')

#vectorize features
assembler = VectorAssembler(inputCols=cols, outputCol='features')
data = assembler.transform(df)

#train/test split
train, test = data.randomSplit([.8, .2])

##Configuring ML pipeline (might use format later)
gbf = GBTRegressor(featuresCol='features', labelCol='sum(GHI)')
pipeline = Pipeline(stages=[gbf])


paramGrid = (ParamGridBuilder()
                 .addGrid(gbf.maxDepth, [5, 10, 15])
                 .build())

crossval = CrossValidator(estimator=pipeline,
                         estimatorParamMaps=paramGrid,
                         evaluator=RegressionEvaluator(labelCol='sum(GHI)'),
                         numFolds=3)

gbfmodel = crossval.fit(train)

gbf1 = gbfmodel.bestModel.stages[0] 
gbf1.save("/home/dyllanjr/Solar_Irradiance_Prediction/models/Gradient_Boosted")

evaluator = RegressionEvaluator(labelCol='sum(GHI)')
train_preds = gbf1.transform(train)
print("Train RMSE: ", evaluator.evaluate(train_preds))
evaluator = RegressionEvaluator(labelCol='sum(GHI)', metricName='r2')
print("Train r2: ", evaluator.evaluate(train_preds))

evaluator = RegressionEvaluator(labelCol='sum(GHI)')
test_preds = gbf1.transform(test)
print("Test RMSE: ", evaluator.evaluate(test_preds))
evaluator = RegressionEvaluator(labelCol='sum(GHI)', metricName='r2')
print("Test r2: ", evaluator.evaluate(test_preds))
